Start supporting a Sync-Read workfload.
Getting the data dynamically is not fun, and it's extremely slow at the
best of times.  So, instead we will implement Sync-* cmdlets that the
user can use for initial data sync, and then use the Get-* cmdlets to
filter and run calculations on the data.

We also open a Sqlite connection and put in a global variable, and use
it persistently.  This counteracts having to carefully use a DataSource
or open connection, which prevents locks.  We can also create and commit
transactions if we're careful (currently I am not, because my catch and
finally clauses are sprinkled with a bunch of shots in the dark to
resolve open transactions; if a transaction remains open and does not
close, it causes a deadlock which requires the user to close the shell.
Luckily finally clauses seem to trigger fairly immediately as the
try-catch scope is exited).

Dumping the data into sqlite also allows us to use other programs (like
R) or import the data into spreadsheets.

Using PoshRSJob speeds up our data captures (currently hardcoding a job
throttle of 100), but it also seems to kill tab completion in the
PowerShell host.  I want to try seeing if different PS hosts will work
to counteract this (like PS6 or PS7).

The user can also change the default sqlite datasource by setting
$global:EvePsDataSource before using Open-EvePsDataConnection.

Also, we split things into more modules to hopefully compartmentalize
stuff a bit.