Some general updates during rewrite into C#.
Got SQLite repository pattern class implemented for category, group,
type, and market group.

Started on the repository pattern class for the ESI endpoints.  I
decided to use a separate repository for the web cache so we're not
mixing which database does what.  While it probably isn't an issue, it
will make it easier if we ever want to move from SQLite to Postgres
later.

Additionally, I decided to use STRICT tables for SQLite.  It caught some
things that were inconsistent (like FLOAT vs DOUBLE vs REAL), though I
do miss being able to explicitly use DATETIME as a column type.

Next on the list will be to do some basic web calls and collect that
data so I can quickly mock the remote endpoint.  I'll also need to keep
in mind how to make the web calls massively threaded, if possible,
because pulling 14k item types will be a chore if I decide to go that
route.